refactor(memory): drop v1 modules; rename v2 modules
Removes v1 RON-backed memory: consolidation.rs, extraction.rs,
scope.rs, the original store.rs + tools.rs, and the v1
memory_integration.rs test file. Renames store_v2.rs → store.rs and
tools_v2.rs → tools.rs to land the v2 module shape.

Updates every crate::ai::memory::store_v2 / tools_v2 reference and
purges the v1 ai/grok/ai_channel integration tests that tested
extraction/consolidation behaviour gone with v2; equivalent coverage
lives in tests/memory_v2.rs (T16).

Revives the 18 previously-ignored tests in ai.rs, grok.rs, and
ai_channel.rs by wiring them through the legacy path (memory.enabled =
false) so the bot sends replies via say_in_reply_to as expected. Deletes
two irrecoverable v1 tests: ai_command_saves_memory_extraction (expected
RON-format ai_memory.ron) and memory_extractor_rejects_web_tool_calls
(expected v1 save_memory/get_memories extractor toolset).

Final test run: 267 passed, 1 skipped (v1_store_discarded — covered by
store::tests::open_renames_v1_store_when_present unit test).
